Colocate chat task tests next to shared task helpers.
Move generateChatName and task view profile tests into src/shared/tasks.test.ts and remove misplaced duplicates.
